Separated trading views from views
